+++ src/components/actions/getApplicationData.ts
@@ -1,8 +1,8 @@
-import { ActionApplicationData, ActionPayload } from '../types'
-import DBConnect from './databaseConnect'
+import { ActionApplicationData, ActionPayload } from '../../types'
+import DBConnect from '../databaseConnect'
 import { BasicObject } from '@openmsupply/expression-evaluator/lib/types'
-import { getAppEntryPointDir } from './utilityFunctions'
-import config from '../config'
+import { getAppEntryPointDir } from '../utilityFunctions'
+import config from '../../config'
 
 // Add more data (such as org/review, etc.) here as required
 export const getApplicationData = async (input: {
