+++ src/components/actions/runAction.ts
@@ -1,0 +1,3 @@
+const routeRunAction = (request: any, reply: any) => {}
+
+export default routeRunAction
