+++ src/server.ts
@@ -17,6 +17,7 @@ import {
   createFilesFolder,
   filesFolder,
 } from './components/files/fileHandler'
+import routeRunAction from './components/actions/runAction'
 import { getAppEntryPointDir } from './components/utilityFunctions'
 import DBConnect from './components/databaseConnect'
 import config from './config'
@@ -61,6 +62,7 @@ const startServer = async () => {
   server.post('/login-org', routeLoginOrg)
   server.get('/updateRowPolicies', routeUpdateRowPolicies)
   server.post('/create-hash', routeCreateHash)
+  server.post('/run-action', routeRunAction)
 
   // File upload endpoint
   server.post('/upload', async function (request: any, reply) {
